feat: migrate RNTester NativeExampleViews to support SceneDelegate
